Added a guard for syntax errors in logger.
Previously, the logger would hang if there was a syntax error in
the command provided. Now, logger will just exit, printing an error
message that the command could not complete.

Note: if we ever wanted to, we could change the behavior to restart the
crashed shell by changing the handler for the KeyboardInterrupt exception
to see if it came from the child thread teeing the shell's stdout/stderr
or from the user hitting Ctrl+C, restarting the shell in the former case
and exiting in the later case.