refactor!: Switch some args to keyword-only
Ruff discovered a number of instances in which boolean arguments were
being used in functions:

* https://docs.astral.sh/ruff/rules/boolean-type-hint-positional-argument/
* https://docs.astral.sh/ruff/rules/boolean-default-value-positional-argument/

The intent all along was that these (along with any others that had
default values associated) be keyword-only arguments, but they were not
denoted as such.  This commit remedies that oversight.

Note that this is a breaking change because it changes the call
signatures of the `Shell` and `ShellLogger` public classes.  To update
your usage, simply switch from positional to keyword arguments for any
arguments affected.