refactor!: Rename trace module
Ruff A005 detected that the `trace` module was shadowing a built-in.  We
should've noticed this years ago.  This commit renames it and the
classes therein so there's no confusion between what's provided by
`shell-logger` vs the standard library.  It also improves the parity
between this module and `stats_collectors`, which improves readability
and eases maintainability.

Note that this is a (rather unfortunate) breaking change.