patch: Add SLSA provenance to release assets
See https://slsa.dev/ for motivation.

Creating a patch release to ensure these additions to the automated
release process work.

Note that the `release` job has been subdivided, because the SLSA
provenance reusable workflow cannot be used as a step within a job, but
must be used as a job on its own.